refactor: adopt hybrid src architecture and align docs
reorganize src into core/reactivity/render/features while keeping public API stable

update internal imports across app, render pipeline, features, tests and demo

refresh README and project docs to reflect the new architecture and boundaries

document render/commit -> features/style exception and update PR template paths
