+++ .github/pull_request_template.md
@@ -20,7 +20,7 @@
 - [ ] `bun test` passes locally (all tests green)
 - [ ] `bun run typecheck` passes locally
 - [ ] Commit messages follow [Conventional Commits](https://www.conventionalcommits.org/) format
-- [ ] Hot path invariant respected: no DOM reads added to `reflow.ts`, `fast-path.ts`, `flex.ts`, or `commit.ts`
+- [ ] Hot path invariant respected: no DOM reads added to `src/render/reflow.ts`, `src/render/engines/fast-path.ts`, `src/render/engines/flex.ts`, or `src/render/commit.ts`
 - [ ] No runtime dependencies added to `src/`
 
 ## Related issues
