+++ src/render/engines/grid.ts
@@ -3,7 +3,7 @@ import type {
   LayoutResult,
   LayoutProps,
   LayoutConstraints,
-} from './types.js'
+} from '../../core/types.js'
 
 import {
   getNodeIndex,
@@ -13,9 +13,9 @@ import {
   getTextHandle,
   getTextContent,
   getLayoutProps,
-} from './prepare.js'
+} from '../prepare.js'
 
-import { resolveResponsiveLayout } from './responsive.js'
+import { resolveResponsiveLayout } from '../strategy/responsive.js'
 import { measureSimple } from './fast-path.js'
 import { measureFlex } from './flex.js'
