+++ tests/edge-cases.test.ts
@@ -1,11 +1,11 @@
 import { describe, test, expect, beforeAll } from 'bun:test'
 import { Window } from 'happy-dom'
-import { defineComponent } from '../src/component.js'
-import { prepare, resetIndexCounter, countNodes } from '../src/prepare.js'
-import { reflow } from '../src/reflow.js'
-import { commitFull } from '../src/commit.js'
-import type { DOMState } from '../src/commit.js'
-import type { LayoutConstraints } from '../src/types.js'
+import { defineComponent } from '../src/render/component.js'
+import { prepare, resetIndexCounter, countNodes } from '../src/render/prepare.js'
+import { reflow } from '../src/render/reflow.js'
+import { commitFull } from '../src/render/commit.js'
+import type { DOMState } from '../src/render/commit.js'
+import type { LayoutConstraints } from '../src/core/types.js'
 
 // ============================================================
 // Setup happy-dom
