+++ tests/portal.test.ts
@@ -1,16 +1,16 @@
 import { describe, test, expect, beforeEach, beforeAll } from 'bun:test'
 import { Window } from 'happy-dom'
-import { createPortal } from '../src/portal.js'
-import { prepare, resetIndexCounter, getNodeIndex, getNodeType, getPreparedChildren, getPortalCssManaged } from '../src/prepare.js'
-import { reflow } from '../src/reflow.js'
-import { commitFull } from '../src/commit.js'
-import type { DOMState } from '../src/commit.js'
-import { defineComponent } from '../src/component.js'
-import { signal } from '../src/signals.js'
+import { createPortal } from '../src/features/portal.js'
+import { prepare, resetIndexCounter, getNodeIndex, getNodeType, getPreparedChildren, getPortalCssManaged } from '../src/render/prepare.js'
+import { reflow } from '../src/render/reflow.js'
+import { commitFull } from '../src/render/commit.js'
+import type { DOMState } from '../src/render/commit.js'
+import { defineComponent } from '../src/render/component.js'
+import { signal } from '../src/reactivity/signals.js'
 import { createApp } from '../src/app.js'
 import { resetScheduler } from '../src/scheduler.js'
-import type { PortalNode, ElementNode } from '../src/types.js'
-import { fullDiff } from '../src/diff.js'
+import type { PortalNode, ElementNode } from '../src/core/types.js'
+import { fullDiff } from '../src/render/diff.js'
 
 // ============================================================
 // DOM setup for commit/app tests
