+++ tests/reflow.test.ts
@@ -1,8 +1,8 @@
 import { describe, test, expect, mock, beforeEach } from 'bun:test'
-import { defineComponent } from '../src/component.js'
-import { prepare, resetIndexCounter } from '../src/prepare.js'
-import { reflow, createLayoutResult } from '../src/reflow.js'
-import type { PreparedComponent } from '../src/types.js'
+import { defineComponent } from '../src/render/component.js'
+import { prepare, resetIndexCounter } from '../src/render/prepare.js'
+import { reflow, createLayoutResult } from '../src/render/reflow.js'
+import type { PreparedComponent } from '../src/core/types.js'
 
 // ============================================================
 // Fake text layout engine
